HBAI-339 Add support for native function calling
