+++ backends/tools/tool.py
@@ -1,19 +1,31 @@
-import os
-import re
 import json
 from enum import Enum
 import importlib.util
 from fastapi import Request
 from typing import Any, Awaitable, List, Optional, Type
 from pydantic import BaseModel
+from tools.helpers import (
+    TOOL_FUNCTION_NAME,
+    find_tool_in_response,
+    get_llm_required_args,
+    get_provided_args,
+    get_required_examples,
+    get_required_schema,
+    import_tool_function,
+    load_function,
+    parse_structured_llm_response,
+    schema_to_markdown,
+    tool_to_json_schema,
+    tool_to_markdown,
+    tool_to_typescript_schema,
+)
 from core import common
-from core.classes import ToolDefinition, ToolFunctionParameter, ToolFunctionSchema
+from core.classes import ToolDefinition, ToolFunctionSchema
 from inference.helpers import KEY_PROMPT_MESSAGE
 from inference.llama_cpp import LLAMA_CPP
 from inference.classes import AgentOutput
 
 # The name to look for when loading python module
-TOOL_FUNCTION_NAME = "main"
 TOOL_NAME = "tool_name"
 example_schema = {TOOL_NAME: "name"}
 TOOL_CHOICE_SCHEMA = f"```json\n{json.dumps(example_schema, indent=4)}\n```"
@@ -28,6 +40,8 @@ class TOOL_SCHEMA_TYPE(str, Enum):
 # @TODO May have better luck coercing Gemma3 using:
 # https://www.reddit.com/r/LocalLLaMA/comments/1jauy8d/giving_native_tool_calling_to_gemma_3_or_really/
 # and https://www.philschmid.de/gemma-function-calling
+# List of Ollama tool models -> https://ollama.com/search?c=tools
+# https://huggingface.co/cfahlgren1/natural-functions -- https://ollama.com/calebfahlgren/natural-functions
 class Tool:
     """
     Handles reading, loading and execution of tool functions and their core deps (if any).
@@ -42,6 +56,7 @@ class Tool:
     def __init__(self, request: Optional[Request] = None):
         self.request = request
         self.filename: str = None
+        # @TODO Dont think we need these here
         self.func_definition: ToolFunctionSchema = None
         self.func: Awaitable[Any] = None
 
@@ -223,9 +238,45 @@ class Tool:
             print(f"{common.PRNT_API} Error loading tool function: {err}", flush=True)
             raise err
 
+    # @TODO Work-in-progress. Difficulty getting Functionary to output args.
+    async def native_call(
+        self,
+        tool_defs: List[ToolDefinition],
+        llm: Type[LLAMA_CPP] = None,
+        query: str = "",
+    ):
+        native_tool_defs = ""
+        for tool_def in tool_defs:
+            # Determine schema format to use for native func calling
+            if llm.tool_schema_type == TOOL_SCHEMA_TYPE.TYPESCRIPT.value:
+                def_str = tool_def.get("typescript_schema", "")
+            else:
+                def_str = tool_def.get("json_schema", "")
+            native_tool_defs += f"\n\n{def_str}"
+        # Prompt the LLM for a response using the tool's schema.
+        # A lower temperature is better for tool use.
+        llm_tool_use_response = await llm.text_completion(
+            prompt=query,
+            system_message="",
+            stream=False,
+            request=self.request,
+            native_tool_defs=native_tool_defs,
+        )
+        content: List[dict] = [item async for item in llm_tool_use_response]
+        # Parse the output
+        data: AgentOutput = content[0].get("data")
+        arguments_response_str = data.get("text")
+        print(
+            f"{common.PRNT_API} Native tool call structured output:\n{arguments_response_str}",
+            flush=True,
+        )
+        parsed_llm_response = json.loads(arguments_response_str)
+        func_call_result = await self.func(**parsed_llm_response)
+        return dict(raw=func_call_result, text=str(func_call_result))
+
     # Execute the tool function with the provided arguments (if any)
     # Return results in raw (for output to other funcs) and string (for text response) formats
-    async def call(
+    async def universal_call(
         self,
         tool_def: ToolDefinition,
         llm: Type[LLAMA_CPP] = None,
@@ -233,8 +284,7 @@ class Tool:
     ) -> AgentOutput:
         self.func_definition = tool_def
         self.func = load_function(tool_def.get("path", None))
-
-        # If llm is not required, skip to func call
+        # Determine if llm is required, if so call the function
         tool_params = tool_def.get("params", None)
         required_llm_arguments = get_llm_required_args(tool_params)
         if len(required_llm_arguments) > 0:
@@ -260,23 +310,12 @@ class Tool:
             tool_example_json = json.dumps(params_example_dict, indent=4)
             tool_example_str = f"```json\n{tool_example_json}\n```"
             tool_args_str = schema_to_markdown(params_schema_dict)
-            # Determine schema format to use for native func calling
-            if llm.tool_schema_type == TOOL_SCHEMA_TYPE.TYPESCRIPT.value:
-                tool_native_args_str = tool_def.get("typescript_schema", "")
-            else:
-                tool_native_args_str = tool_def.get("json_schema", "")
             # Inject template args into prompt template (Universal func calling)
-            tool_prompt = query
-            if not llm.is_tool_capable:
-                tool_prompt = universal_tool_template.replace(KEY_PROMPT_MESSAGE, query)
-                tool_prompt = tool_prompt.replace(TOOL_ARGUMENTS, tool_args_str)
-                tool_prompt = tool_prompt.replace(
-                    TOOL_EXAMPLE_ARGUMENTS, tool_example_str
-                )
-                tool_prompt = tool_prompt.replace(TOOL_NAME, tool_name_str)
-                tool_prompt = tool_prompt.replace(
-                    TOOL_DESCRIPTION, tool_description_str
-                )
+            tool_prompt = universal_tool_template.replace(KEY_PROMPT_MESSAGE, query)
+            tool_prompt = tool_prompt.replace(TOOL_ARGUMENTS, tool_args_str)
+            tool_prompt = tool_prompt.replace(TOOL_EXAMPLE_ARGUMENTS, tool_example_str)
+            tool_prompt = tool_prompt.replace(TOOL_NAME, tool_name_str)
+            tool_prompt = tool_prompt.replace(TOOL_DESCRIPTION, tool_description_str)
             # Prompt the LLM for a response using the tool's schema.
             # A lower temperature is better for tool use.
             llm_tool_use_response = await llm.text_completion(
@@ -284,14 +323,13 @@ class Tool:
                 system_message=system_message,
                 stream=False,
                 request=self.request,
-                native_tool_defs=tool_native_args_str,  # for native func calling
             )
             content: List[dict] = [item async for item in llm_tool_use_response]
             data: AgentOutput = content[0].get("data")
             # Parse out the json result using regex
             arguments_response_str = data.get("text")
             print(
-                f"{common.PRNT_API} Tool call structured output:\n{arguments_response_str}",
+                f"{common.PRNT_API} Universal tool call structured output:\n{arguments_response_str}",
                 flush=True,
             )
             parsed_llm_response = parse_structured_llm_response(
@@ -305,237 +343,9 @@ class Tool:
             )
             func_call_result = await self.func(**parsed_llm_response)
             return dict(raw=func_call_result, text=str(func_call_result))
-        # Call function with arguments provided by the tool and/or prompt
-        func_arguments = get_provided_args(prompt=query, tool_params=tool_params)
-        func_call_result = await self.func(**func_arguments)
-        # Return results
-        return dict(raw=func_call_result, text=str(func_call_result))
-
-
-def import_tool_function(filename: str):
-    spec = None
-
-    # Check built-in funcs first
-    try:
-        # from _deps directory
-        prebuilt_funcs_path = common.dep_path(
-            os.path.join(
-                common.BACKENDS_FOLDER,
-                common.TOOL_FOLDER,
-                common.TOOL_FUNCS_FOLDER,
-                filename,
-            )
-        )
-        if os.path.exists(prebuilt_funcs_path):
-            spec = importlib.util.spec_from_file_location(
-                name=filename,
-                location=prebuilt_funcs_path,
-            )
-    except Exception as err:
-        print(f"{common.PRNT_API} {err}", flush=True)
-
-    # Check user made funcs
-    try:
-        # from root of installation dir
-        custom_funcs_path = os.path.join(os.getcwd(), common.TOOL_FUNCS_PATH, filename)
-        if os.path.exists(custom_funcs_path):
-            spec = importlib.util.spec_from_file_location(
-                name=filename,
-                location=custom_funcs_path,
-            )
-    except Exception as err:
-        print(f"{common.PRNT_API} {err}", flush=True)
-
-    if not spec:
-        raise Exception("No tool found.")
-
-    return spec
-
-
-def load_function(filename: str):
-    try:
-        spec = import_tool_function(filename)
-        tool_code = importlib.util.module_from_spec(spec)
-        spec.loader.exec_module(tool_code)
-        func = getattr(tool_code, TOOL_FUNCTION_NAME)
-        return func
-    except Exception as err:
-        print(f"{common.PRNT_API} {err}", flush=True)
-        raise Exception("Failed to load tool function.")
-
-
-# Good for explaining each tool to llm so it can make a selection.
-def tool_to_markdown(tool_list: List[ToolDefinition], include_code=False):
-    """Create a machine readable description of a tool definition."""
-    markdown_string = ""
-    for index, item in enumerate(tool_list):
-        markdown_string += f"## Tool {index + 1}\n\n"
-        for key, value in item.items():
-            # If code
-            if key == "params" or key == "params_example" or key == "params_schema":
-                if include_code:
-                    markdown_string += (
-                        f"### {key}\n```json\n{json.dumps(value, indent=4)}\n```\n\n"
-                    )
-            # Add descr
-            elif key == "description" or key == "name":
-                markdown_string += f"### {key}\n\n{value}\n\n"
-            # Add tool return type
-            elif key == "output_type" and include_code:
-                return_type = ", ".join(value)
-                if return_type:
-                    markdown_string += f"### Return type:\n\n{return_type}"
-    return markdown_string
-
-
-# Parse out the json from llm tool call response using either regex or another llm call
-def parse_structured_llm_response(
-    arguments_str: str, allowed_arguments: List[str]
-) -> dict:
-    pattern_object = r"({.*?})"
-    pattern_json_object = r"\`\`\`json\n({.*?})\n\`\`\`"
-    match_json_object = re.search(pattern_json_object, arguments_str, re.DOTALL)
-    match_object = re.search(pattern_object, arguments_str, re.DOTALL)
-
-    if match_json_object or match_object:
-        # Find first occurance
-        if match_json_object:
-            json_block = match_json_object.group(1)
-        elif match_object:
-            json_block = match_object.group(1)
-        # Remove single-line comments (//...)
-        json_block = re.sub(r"//.*", "", json_block)
-        # Remove multi-line comments (/*...*/)
-        json_block = re.sub(r"/\*.*?\*/", "", json_block, flags=re.DOTALL)
-        # Clean up any extra commas or trailing whitespace
-        json_block = re.sub(r",\s*(\}|\])", r"\1", json_block)
-        json_block = json_block.strip()
-        # Convert JSON block back to a dictionary to ensure it's valid JSON
-        try:
-            # Remove any unrelated keys from json
-            json_object: dict = json.loads(json_block)
-            # Filter out keys not in the allowed_keys set
-            filtered_json_object = {
-                k: v for k, v in json_object.items() if k in allowed_arguments
-            }
-            return filtered_json_object
-        except json.JSONDecodeError as e:
-            raise Exception("Invalid JSON.")
-    else:
-        raise Exception("No JSON block found!")
-
-
-# Conversion func to translate our tool def to a native tool (openai compatible) json schema.
-# Most models will likely use this, but some, like Functionary use other formats.
-def tool_to_json_schema(name: str, description: str, params: List[dict]) -> str:
-    # Construct parameters
-    properties = {}
-    required = []
-    for param in params:
-        param_name = param["name"]
-        param_type = param["type"]
-        param_description = param["description"]
-        param_schema = {"type": param_type, "description": param_description}
-        properties[param_name] = param_schema
-        required.append(param_name)
-    # Construct new schema
-    # @TODO Add output type
-    native_schema = {
-        "type": "function",
-        "function": {
-            "name": name,
-            "description": description,
-            "parameters": {
-                "type": "object",  # For now, props all passed as object
-                "properties": properties,
-                "required": required,
-            },
-        },
-    }
-    return json.dumps(native_schema)
-
-
-# Conversion func to translate our tool def to a native tool (typescript) schema
-def tool_to_typescript_schema(name: str, description: str, params: List[dict]) -> str:
-    # Construct parameters
-    properties = ""
-    output_type = "any"  # @TODO Define output type
-    for param in params:
-        param_name = param["name"]
-        param_type = param["type"]
-        param_description = param["description"]
-        param_str = f"// {param_description}\n{param_name}: {param_type},"
-        properties += f"\n{param_str}"
-    # Construct new schema
-    native_schema = f"""// {description}
-type {name} = (_: {{
-{properties}
-}}) => {output_type};"""
-    # Result
-    return native_schema
-
-
-# Convert a tool's schema to markdown text
-def schema_to_markdown(schema: dict) -> str:
-    result = ""
-    for index, pname in enumerate(schema):
-        descr = schema[pname].get("description", None)
-        data_type = schema[pname].get("type", None)
-        allowed_values = schema[pname].get("allowed_values", None)
-        if index > 0:
-            result += "\n\n"
-        result += f"### {pname}\n\nDescription: {descr}\nData type: {data_type}"
-        if allowed_values:
-            result += f"\nAllowed values: {allowed_values}"
-    return result
-
-
-def get_required_examples(required: List[str], example: dict) -> dict:
-    result = dict()
-    for pname in required:
-        if pname in example:
-            result[pname] = example[pname]
-    return result
-
-
-def get_required_schema(required: List[str], schema: dict) -> dict:
-    result = dict()
-    for pname in required:
-        if pname in schema:
-            result[pname] = schema[pname]
-    return result
-
-
-def get_provided_args(prompt: str, tool_params: dict):
-    provided_arguments = dict()
-    for pname in tool_params:
-        value = tool_params[pname].get("value", None)
-        if pname == "prompt":
-            provided_arguments[pname] = prompt
-            continue
-        if value:
-            provided_arguments[pname] = value
-    return provided_arguments
-
-
-# Determine allowed arg names (arguments that llm needs to fill in)
-def get_llm_required_args(tool_params: List[ToolFunctionParameter]) -> List[str]:
-    result = []
-    # Check each `value` exists, if not then the llm needs to send it,
-    # we ignore "prompt" since its always provided.
-    for param in tool_params:
-        pname = param.get("name", None)
-        if pname and pname != "prompt" and not param.get("value", None):
-            result.append(pname)
-    return result
-
-
-def find_tool_in_response(response: str, tools: List[str]) -> Optional[str]:
-    """Attempt to find any tool names in the given response text."""
-    print(f"{common.PRNT_API} Searching in response for tools...{tools}", flush=True)
-    for tool_name in tools:
-        match = re.search(re.escape(tool_name), response, re.DOTALL)
-        if match:
-            print(f"{common.PRNT_API} Found tool:{tool_name}", flush=True)
-            return tool_name  # Return first match immediately
-    return None  # Explicitly return None if no tool is found
+        else:
+            # Call function with arguments provided by the tool and/or prompt
+            func_arguments = get_provided_args(prompt=query, tool_params=tool_params)
+            func_call_result = await self.func(**func_arguments)
+            # Return results
+            return dict(raw=func_call_result, text=str(func_call_result))
