+++ public/prompt_formats.json
@@ -80,12 +80,16 @@
     "user": "<|im_start|>user\n{{user_message}}<|im_end|>\n<|im_start|>assistant\n"
   },
   "functionaryV3": {
-    "system": "<|start_header_id|>system<|end_header_id|>\n\nYou are capable of executing available function(s) if required.\nOnly execute function(s) when absolutely necessary.\nAsk for the required input to:recipient==all\nUse JSON for function arguments.\nRespond in this format:\n>>>${recipient}\n${content}\nAvailable functions:\n// Supported function definitions that should be called when necessary.\nnamespace functions {\n\n{{tool_defs}}\n\n} // namespace functions<|eot_id|>",
-    "user": "<|start_header_id|>user<|end_header_id|>\n\n{{user_message}}"
+    "system": "<|start_header_id|>system<|end_header_id|>\n\nYou are capable of executing available function(s) if required.\nOnly execute function(s) when absolutely necessary.\nAsk for the required input to:recipient==all\nUse JSON for function arguments.\nRespond in this format:\n>>>${recipient}\n${content}\nAvailable functions:\n// Supported function definitions that should be called when necessary.\nnamespace functions {{{tool_defs}}\n\n} // namespace functions<|eot_id|>",
+    "user": "<|start_header_id|>user<|end_header_id|>\n\n{{user_message}}<|eot_id|><|start_header_id|>assistant<|end_header_id|>"
   },
   "functionaryV2": {
-    "system": "<|from|>system\n<|recipient|>all\n<|content|>// Supported function definitions that should be called when necessary.\nnamespace functions {\n\n{{tool_defs}}\n\n} // namespace functions\n<|from|>system\n<|recipient|>all\n<|content|>{{system_message}}\n",
-    "user": "<|from|>user\n<|recipient|>all\n<|content|>{{user_message}}"
+    "system": "<|from|>system\n<|recipient|>all\n<|content|>// Supported function definitions that should be called when necessary.\nnamespace functions {{{tool_defs}}\n\n} // namespace functions\n<|from|>system\n<|recipient|>all\n<|content|>{{system_message}}\n",
+    "user": "<|from|>user\n<|recipient|>all\n<|content|>{{user_message}}<|from|>assistant\n<|recipient|>all<|content|>"
+  },
+  "hammer2": {
+    "system": "<|im_start|>system\n{{system_message}}<|im_end|>\n",
+    "user": "<|im_start|>user\n[BEGIN OF TASK INSTRUCTION]\nYou are a tool calling assistant. In order to complete the user's request, you need to select one or more appropriate tools from the following tools and fill in the correct values for the tool parameters. Your specific tasks are:\n1. Make one or more function/tool calls to meet the request based on the question.\n2. If none of the function can be used, point it out and refuse to answer.\n3. If the given question lacks the parameters required by the function, also point it out.\n\nThe following are characters that may interact with you\n1. user: Provides query or additional information.\n2. tool: Returns the results of the tool calling.\n[END OF TASK INSTRUCTION]\n\n[BEGIN OF TASK INSTRUCTION]\n\nThe output MUST strictly adhere to the following JSON format, and NO other text MUST be included.\nThe example format is as follows. Please make sure the parameter type is correct. If no function call is needed, please directly output an empty list '[]'\n```\n[\n    {\"name\": \"func_name1\", \"arguments\": {\"argument1\": \"value1\", \"argument2\": \"value2\"}},\n    ... (more tool calls as required)\n]\n[END OF TASK INSTRUCTION]\n\n<|im_end|>\n[BEGIN OF AVAILABLE_TOOLS]\n{{{tool_defs}}\n[END OF AVAILABLE_TOOLS]\n\n<|im_start|>user\n{{user_message}}<|im_end|>\n<|im_start|>assistant\n"
   },
   "qwen": {
     "system": "<|im_start|>system\n{{system_message}}<|im_end|>\n",
